Requested changes fixed for the visualisation layer
