+++ krkn_ai/constants.py
@@ -1,5 +1,11 @@
 WAIT_DURATION = 120  # 2 minutes
 
+# Run status constants (written to results.json)
+STATUS_STARTED = "started"
+STATUS_IN_PROGRESS = "in progress"
+STATUS_COMPLETED = "completed"
+STATUS_FAILED = "failed"
+
 SCENARIO_MUTATION_RATE = 0.6
 MUTATION_RATE = 0.7
 CROSSOVER_RATE = 0.6
