+++ krkn_ai/reporter/json_summary_reporter.py
@@ -10,6 +10,7 @@ from typing import Any, Dict, List, Optional
 from krkn_ai.models.app import CommandRunResult
 from krkn_ai.models.config import ConfigFile
 from krkn_ai.utils.logger import get_logger
+from krkn_ai.constants import STATUS_COMPLETED
 
 logger = get_logger(__name__)
 
@@ -56,7 +57,7 @@ class JSONSummaryReporter:
         self.end_time = end_time
         self.completed_generations = completed_generations
         self.seed = seed
-        self.status = "completed"
+        self.status = STATUS_COMPLETED
 
     def generate_summary(self) -> Dict[str, Any]:
         """
